test-common/requirements.txt
cassandra-driver
slackclient

